pyproject.toml
[project]
name = "librarian"
version = "0.1.0"
description = "Librarian is an AI Skills management system built on LangGraph/LangChain. It automates the organization, deduplication, evaluation, and maintenance of AI skills with support for human review workflows, scheduled jobs, and REST API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "langchain>=1.2.15",
    "langchain-openai>=1.2.0",
    "langgraph>=1.1.9",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]
